output JSON log incrementally

+++ log.h
@@ -19,8 +19,8 @@
 
 typedef struct log_save_error
 {
-    gint wrote;
-    gint length;
+    size_t wrote;
+    size_t length;
 } log_save_error_t;
 
 void log_open(GSList*, gboolean);
